refactor: separate EQ response graph component
